Break out all reformatting functions.
